Fix the Inception model to work
The current `Inception` model fails to load the published pretrained
model and produce correct predictions. This pull-request will fix this
issue.

One of the causes of the problem lies in the `Inception.MakeBranch()`
method.  The corresponding `make_branch()` method in python can accept
multiple `conv_settings` parameters and is called as follows:

```python
def _make_A(pool_features):
    out = nn.HybridConcurrent(axis=1)
    out.add(_make_branch(None,
                         (64, 1, None, None))) # conv setting 1
    out.add(_make_branch(None,
                         (48, 1, None, None),  # conv setting 1
                         (64, 5, None, 2)))    # conv setting 2
:
```

On the other hand, our `Inception.MakeBranch()` takes only one
`conv_setting` and instead of that, it is called in multiple times as
follows:

```csharp
internal static HybridConcurrent MakeA(int pool_features, string prefix = "")
{
    var output = new HybridConcurrent(1, prefix);
    output.Add(MakeBranch("", 64, (1, 1)));
    output.Add(MakeBranch("", 48, (1, 1)));
    output.Add(MakeBranch("", 64, (5, 5), null, (2, 2)));
    :
```

In the python version, the generated network structure is as follows:

```
HybridConcurrent
   +--HybridSequential
   |     conv2D(channels=64, kernel=(1,1))
   +--HybridSequential
         conv2D(channels=48, kernel=(1,1))
         conv2D(channels=64, kernel=(5,5), padding=(2,2))
:
```

In the C# version, the generated network structure is as follows:

```
HybridConcurrent
   +--HybridSequential
   |     conv2D(channels=64, kernel=(1,1))
   +--HybridSequential
   |     conv2D(channels=48, kernel=(1,1))
   +--HybridSequential
         conv2D(channels=64, kernel=(5,5), padding=(2,2))
:
```

These networks have different semantics because `HybridConcurrent` is a
block that concatenates its children's outputs to a single tensor.

This PR will rewrite `Inception.MakeBranch()` and its call sites to
accept and pass multiple `conv_setting` at once and generate the same
network structure as that of the python implementation.

The second type of changes is to introduce name scope declarations to
the `Inception` class. Name scopes don't affect network behavior, but
add prefixes to network parameter names. (For example, a parameter name
`conv0` becomes `A_conv0` when it is enclosed with a name scope `A_`.)

To load the pretrained Inception model correctly, we should use name
scopes to obtain compatible parameter names with the pretrained model.

To do so, this PR will enclose each code block with the same `using
(output.NameScope.With())` clause as the `with out.name_scope` clause
in the python implementation.

By the above two types of changes, I made sure that the `Inception`
model works correctly.